requirements.txt
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
tqdm>=4.66
scikit-learn>=1.3
scikit-image>=0.22
lightgbm>=4.1
matplotlib>=3.8
seaborn>=0.13
Pillow>=10.0
joblib>=1.3
torch>=2.2
torchvision>=0.17
